Slight Optimizations
Added slight optimization by passing in remaining_words as a parameter to Grid instead of recalculating it every time. Ideas to further optimize by removing extract_words and storing just the coordinates as instance variable (and just query the word at a desired position when needed). Note: contact jhu04 if want to edit this as it is filled with debug statements and spaghetti code